Address WASM PInvoke review feedback
