+++ src/tests/Interop/PInvoke/Unresolved/UnresolvedPInvokeTests.csproj
@@ -1,0 +1,8 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <RequiresProcessIsolation>true</RequiresProcessIsolation>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="UnresolvedPInvokeTests.cs" />
+  </ItemGroup>
+</Project>
