+++ src/tests/JIT/Regression_o_2/Runtime_58874.cs
@@ -12,8 +12,6 @@ using Xunit;
 
 public unsafe class Runtime_58874
 {
-    private static bool s_enteredMissingPInvokeMethod;
-
     [Fact]
     public static void TestEntryPoint()
     {
@@ -23,25 +21,6 @@ public unsafe class Runtime_58874
             Foo(endOfPage.Pointer);
         }
     }
-
-    [Fact]
-    public static void MissingPInvokeThrowsWhenCalled()
-    {
-        s_enteredMissingPInvokeMethod = false;
-
-        Assert.Throws<DllNotFoundException>(CallMissingPInvoke);
-        Assert.True(s_enteredMissingPInvokeMethod);
-    }
-
-    [MethodImpl(MethodImplOptions.NoInlining)]
-    private static void CallMissingPInvoke()
-    {
-        s_enteredMissingPInvokeMethod = true;
-        MissingPInvoke();
-    }
-
-    [DllImport("Runtime_58874_MissingNativeLibrary")]
-    private static extern void MissingPInvoke();
 
     [MethodImpl(MethodImplOptions.NoInlining)]
     private static Test Foo(Test* t)
